# spec/openapi/nvc-practice-coach.v1.yaml
openapi: 3.1.0
info:
  title: NVC Practice Coach API
  version: 1.0.0
  description: API contract for the NVC Practice Coach MVP.
servers:
  - url: https://api.nvc-coach.example.com/api/v1
    description: Production
  - url: http://localhost:3000/api/v1
    description: Local
security:
  - bearerAuth: []
...
paths:
...
  /sessions:
    get:
      tags: [sessions]
      operationId: listSessions
      summary: Get current user's session history list
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
        - in: query
          name: offset
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: state
          required: false
          schema:
            type: string
            enum: [ACTIVE, COMPLETED, ABANDONED]
          description: Filter by session state
        - in: query
          name: keyword
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 80
          description: Keyword search in scene title/goal/context and latest turn messages
        - in: query
          name: created_from
          required: false
          schema:
            type: string
            format: date
          description: Inclusive start date (YYYY-MM-DD)
        - in: query
          name: created_to
          required: false
          schema:
            type: string
            format: date
          description: Inclusive end date (YYYY-MM-DD)
      responses:
        '200':
          description: Session history list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionHistoryListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [sessions]
      operationId: createSession
      summary: Start a practice session from an existing scene
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionCreateRequest'
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionCreateResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /voice/sessions:
    post:
      tags: [voice]
      operationId: startVoiceSession
      summary: Start an Agora voice practice session from an existing scene
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoiceSessionStartRequest'
      responses:
        '201':
          description: Voice session started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoiceSessionStartResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalError'
  /voice/sessions/{session_id}/stop:
    post:
      tags: [voice]
      operationId: stopVoiceSession
      summary: Stop a voice practice session by persisted Agora agent id
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Voice session stopped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoiceSessionStopResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
  /voice/sessions/{session_id}/transcripts:
    post:
      tags: [voice]
      operationId: syncVoiceTranscripts
      summary: Persist stable voice transcript turns and generate OFNR feedback
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoiceTranscriptSyncRequest'
      responses:
        '200':
          description: Transcript turns synced
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoiceTranscriptSyncResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SessionId:
      in: path
      name: session_id
      required: true
      schema:
        type: string
        format: uuid
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: VALIDATION_ERROR
            message: goal is required
            request_id: f2cf1f5e-3d4d-4f1f-bffc-e103a7f03a8a
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: UNAUTHORIZED
            message: missing or invalid token
            request_id: 4f0e9d9c-f54c-4098-96ca-9d6f06f725fd
    NotFoundError:
      description: Requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: NOT_FOUND
            message: session not found
            request_id: dc6a22cb-2892-4e9c-b678-e4605a9f11df
    ConflictError:
      description: Resource state conflicts with request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: CONFLICT
            message: reflection already exists for this session
            request_id: a3fec497-7747-4209-86bb-499b1608c3cc
...
    InternalError:
      description: Unexpected internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: INTERNAL_ERROR
            message: internal server error
            request_id: 5bf1a3a9-1dc8-4128-bff1-455713e6a0dc
  schemas:
...
    SessionState:
      type: string
      enum: [ACTIVE, COMPLETED, ABANDONED]
    SessionModality:
      type: string
      enum: [TEXT, VOICE]
...
    RiskLevel:
      type: string
      enum: [LOW, MEDIUM, HIGH]
...
    SessionCreateRequest:
      type: object
      additionalProperties: false
      required: [scene_id, target_turns]
      properties:
        scene_id:
          type: string
          format: uuid
        target_turns:
          type: integer
          minimum: 5
          maximum: 8
    SessionCreateResponse:
      type: object
      additionalProperties: false
      required: [session_id, state, current_turn, created_at]
      properties:
        session_id:
          type: string
          format: uuid
        state:
          $ref: '#/components/schemas/SessionState'
        current_turn:
          type: integer
          minimum: 0
        created_at:
          type: string
          format: date-time
    SessionHistoryListItem:
      type: object
      additionalProperties: false
      required:
        - session_id
        - scene_id
        - scene_title
        - state
        - modality
        - current_turn
        - target_turns
        - created_at
        - has_summary
        - has_reflection
      properties:
        session_id:
          type: string
          format: uuid
        scene_id:
          type: string
          format: uuid
        scene_title:
          type: string
        state:
          $ref: '#/components/schemas/SessionState'
        modality:
          $ref: '#/components/schemas/SessionModality'
          default: TEXT
        current_turn:
          type: integer
          minimum: 0
        target_turns:
          type: integer
          minimum: 5
          maximum: 8
        created_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
        last_user_message:
          type: string
        last_assistant_message:
          type: string
        last_overall_score:
          type: integer
          minimum: 0
          maximum: 100
        last_risk_level:
          $ref: '#/components/schemas/RiskLevel'
        has_summary:
          type: boolean
        has_reflection:
          type: boolean
    SessionHistoryListResponse:
      type: object
      additionalProperties: false
      required: [items, limit, offset, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SessionHistoryListItem'
        limit:
          type: integer
          minimum: 1
          maximum: 50
        offset:
          type: integer
          minimum: 0
        total:
          type: integer
          minimum: 0
...
    VoiceSessionStartRequest:
      type: object
      additionalProperties: false
      required: [scene_id]
      properties:
        scene_id:
          type: string
          format: uuid
        target_turns:
          type: integer
          minimum: 5
          maximum: 8
          default: 6
    VoiceSessionStartResponse:
      type: object
      additionalProperties: false
      required:
        - session_id
        - app_id
        - token
        - uid
        - channel_name
        - agent_uid
        - agent_id
        - expires_at
      properties:
        session_id:
          type: string
          format: uuid
        app_id:
          type: string
        token:
          type: string
        uid:
          type: string
        channel_name:
          type: string
        agent_uid:
          type: string
        agent_id:
          type: string
        expires_at:
          type: string
          format: date-time
    VoiceSessionStopResponse:
      type: object
      additionalProperties: false
      required: [session_id, status]
      properties:
        session_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [STOPPED]
    VoiceTranscriptTurn:
      type: object
      additionalProperties: false
      required: [role, content, external_turn_id]
      properties:
        role:
          type: string
          enum: [USER, ASSISTANT]
        content:
          type: string
          minLength: 1
          maxLength: 4000
        external_turn_id:
          type: string
          minLength: 1
          maxLength: 120
        metadata:
          type: object
          additionalProperties: true
          default: {}
    VoiceTranscriptSyncRequest:
      type: object
      additionalProperties: false
      required: [turns]
      properties:
        turns:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/VoiceTranscriptTurn'
    VoiceTranscriptSyncResponse:
      type: object
      additionalProperties: false
      required: [session_id, inserted_count, current_turn, state]
      properties:
        session_id:
          type: string
          format: uuid
        inserted_count:
          type: integer
          minimum: 0
        current_turn:
          type: integer
          minimum: 0
        state:
          $ref: '#/components/schemas/SessionState'
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error_code, message, request_id]
      properties:
        error_code:
          type: string
          enum:
            - VALIDATION_ERROR
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - SAFETY_BLOCKED
            - RATE_LIMITED
            - INTERNAL_ERROR
        message:
          type: string
        request_id:
          type: string
          format: uuid